refactor(app): extract session history window controller
